Split into variants for "play" and "fdroid"
play: Works exactly as before.
fdroid: does not contain any ads/ad libraries making the app compatible
  with FDroid again.

Both variants share the same application ID so users of both platforms
will get seamless upgrades.

Fixes #48.
