+++ app/build.gradle
@@ -24,6 +24,19 @@ android {
         disable "ResourceType"
         disable 'GoogleAppIndexingWarning'
     }
+
+    flavorDimensions "distribution"
+
+    productFlavors {
+        play {
+            dimension "distribution"
+        }
+
+        fdroid {
+            dimension "distribution"
+            versionNameSuffix "-fdroid"
+        }
+    }
 }
 
 dependencies {
@@ -32,9 +45,10 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'androidx.emoji:emoji-appcompat:1.1.0'
     implementation 'androidx.preference:preference-ktx:1.1.1'
-    implementation 'com.google.android.gms:play-services-ads-lite:19.7.0'
     implementation 'com.mobeta.android.dslv:library:0.9.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    playImplementation 'com.google.android.gms:play-services-ads-lite:19.7.0'
 }
 repositories {
     mavenCentral()
