+++ app/src/fdroid/java/jp/ddo/hotmist/unicodepad/AdCompatImpl.kt
@@ -1,0 +1,10 @@
+package jp.ddo.hotmist.unicodepad
+
+import android.app.Activity
+import android.content.SharedPreferences
+
+internal class AdCompatImpl : AdCompat {
+    override val showAdSettings = false
+    override fun renderAdToContainer(activity: Activity, pref: SharedPreferences) {
+    }
+}
