+++ app/src/fdroid/res/layout/ad_container.xml
@@ -1,0 +1,1 @@
+<merge xmlns:android="http://schemas.android.com/apk/res/android" />
