+++ app/src/play/res/layout/ad_container.xml
@@ -1,0 +1,8 @@
+<LinearLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+
+        android:id="@+id/adContainer"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+</LinearLayout>
